abstract function getdonationurl
