+++ src/features/projects/components/PopupProject.tsx
@@ -1,14 +1,11 @@
-import Modal from '@material-ui/core/Modal';
 import React, { ReactElement, Ref } from 'react';
 import getImageUrl from '../../../utils/getImageURL';
-import { ThemeContext } from '../../../theme/themeContext';
-import DonationsPopup from '../../donations';
 import i18next from '../../../../i18n/'
 import getFormatedCurrency from '../../../utils/countryCurrency/getFormattedCurrency';
 import { localizedAbbreviatedNumber } from '../../../utils/getFormattedNumber';
 import { truncateString } from '../../../utils/getTruncatedString';
-import getStoredCurrency from '../../../utils/countryCurrency/getStoredCurrency';
 import { UserPropsContext } from '../../common/Layout/UserPropsContext';
+import { getDonationUrl } from '../../../utils/getDonationUrl';
 
 const { useTranslation } = i18next;
 interface Props {
@@ -29,8 +26,7 @@ export default function PopupProject({
   popupRef,
 }: Props): ReactElement {
   const { t, i18n, ready } = useTranslation(['donate', 'common', 'country']);
-  const { theme } = React.useContext(ThemeContext);
-  const { user } = React.useContext(UserPropsContext);
+  const { token } = React.useContext(UserPropsContext);
 
 
   const ImageSource = project.properties.image
@@ -42,18 +38,8 @@ export default function PopupProject({
 
   const projectDetails = project.properties;
 
-  const currency = getStoredCurrency();
-  const country = localStorage.getItem('countryCode');
-  const language = localStorage.getItem('language');
-
-  const getSourceUrl = React.useCallback((): string => {
-    let sourceUrl = `${process.env.NEXT_PUBLIC_DONATION_URL}/?to=${projectDetails.slug}&returnToUrl=${window.location.href}&country=${country}&currency=${currency}&locale=${language}${user ? '&autoLogin=true' : ''}&tenant=${process.env.TENANTID}`;
-    return sourceUrl;
-  }, [project, country, currency, language, user]);
-
-  const url = getSourceUrl();
-
   const handleDonationOpen = () => {
+    const url = getDonationUrl(project.properties.slug, token);
     window.location.href = url;
   };
